requirements.txt
pyTelegramBotAPI
selenium
python-dotenv
